Refactoring of the project organisation
