+++ README.md
@@ -1,0 +1,16 @@
+# Calculadora en JavaScript
+
+Esta es una simple calculadora construida con JavaScript plano.
+
+## Instalación
+
+1. Clona este repositorio.
+2. Ejecuta `npm install` para instalar las dependencias.
+
+## Uso
+
+- Ejecuta `npm start` para iniciar el servidor de desarrollo.
+- Abre tu navegador y visita `http://localhost:8080` para ver la calculadora.
+
+## Pruebas
+Ejecuta `npm test` para ejecutar las pruebas con Cucumber.js.
