+++ index.html
@@ -1,10 +1,15 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="es">
 <head>
-    <meta charset="UTF-8">
     <title>Calculadora</title>
+    <link rel="stylesheet" href="styles/styles.css">
 </head>
 <body>
+<div id="calculator">
+    hello world
+</div>
 
+<script src="scripts/calculator.js"></script>
+<script src="scripts/ui.js"></script>
 </body>
 </html>
