+++ package.json
@@ -1,15 +1,26 @@
 {
-  "name": "hellocucumber",
+  "name": "calculator",
   "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "description": "A simple calculator application built with vanilla JavaScript",
+  "keywords": [
+    "calculator",
+    "javascript",
+    "vanilla"
+  ],
   "scripts": {
-    "test": "cucumber-js"
+    "test": "cucumber-js",
+    "start": "live-server",
+    "reset": "rm -rf node_modules && npm install"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "devDependencies": {
-    "@cucumber/cucumber": "^10.7.0"
-  }
+    "@cucumber/cucumber": "^10.8.0",
+    "live-server": "^1.2.2"
+  },
+  "author": "Tu Nombre o Organización",
+  "license": "MIT",
+  "engines": {
+    "node": ">=14.0.0"
+  },
+  "homepage": "https://github.com/jnperezmesa/calculator_js",
+  "bugs": "https://github.com/jnperezmesa/calculator_js/issues"
 }
